Extract constructor chain samples
